First pass at breaking out verifiable_credential_1_1.rs
